+++ pyespn/classes/team.py
@@ -5,6 +5,7 @@ from pyespn.classes.image import Image
 from pyespn.classes.stat import Record, Stat
 from pyespn.data.version import espn_api_version as v
 from pyespn.core.decorators import validate_json
+from pyespn.exceptions import API400Error
 from concurrent.futures import ThreadPoolExecutor, as_completed
 
 
@@ -67,6 +68,7 @@ class Team:
         self.records = {}
         self.stats = {}
         self.coaches = {}
+        self.betting = {}
         if team_json:
             self.team_json = team_json
         else:
@@ -110,15 +112,48 @@ class Team:
         self.links = {link["rel"][0]: link["href"] for link in self.team_json.get("links", []) if "rel" in link}
 
     def load_team_season_stats(self, season):
-        api_info = lookup_league_api_info(league_abbv=self.espn_instance.league_abbv)
+        """
+        Fetches and loads team-level statistical data for a specific season, using threading to fetch multiple pages concurrently.
 
+        This method sends a request to the ESPN API to retrieve team statistics for the given season. It will send requests for
+        multiple pages of data if necessary, and each page is fetched concurrently using a thread pool. The method parses the
+        response and instantiates `Stat` objects for each statistic found under the `categories` section of the response.
+        All parsed statistics are then stored in the `self.stats` dictionary, keyed by the provided season.
+
+        Args:
+            season (int): The season year for which statistics should be retrieved.
+
+        Raises:
+            API400Error: If the API responds with a 400-level error. A message is printed, and no data is stored for the season.
+        """
+        api_info = lookup_league_api_info(league_abbv=self.espn_instance.league_abbv)
         url = f'http://sports.core.api.espn.com/{v}/sports/{api_info["sport"]}/leagues/{api_info["league"]}/seasons/{season}/types/2/teams/{self.team_id}/statistics?lang=en&region=us'
-        stats_content = fetch_espn_data(url)
+
         all_stats = []
-        for stat in stats_content.get('splits', []):
-            all_stats.append(Stat(stat_json=stat,
-                                  espn_instance=self.espn_instance))
-        self.stats[season] = {}
+        try:
+            stats_content = fetch_espn_data(url)
+            pages = stats_content.get('pageCount', 0)
+
+            # Using ThreadPoolExecutor to fetch multiple pages concurrently
+            with ThreadPoolExecutor() as executor:
+                future_to_page = {
+                    executor.submit(fetch_espn_data, f'{url}?page={page}'): page
+                    for page in range(1, pages + 1)
+                }
+
+                # Collect results as they complete
+                for future in as_completed(future_to_page):
+                    page_data = future.result()
+                    for categories in page_data.get('splits', {}).get('categories', []):
+                        for stat in categories.get('stats', []):
+                            all_stats.append(Stat(stat_json=stat,
+                                                  espn_instance=self.espn_instance))
+
+        except API400Error as e:
+            print(f"Failed to fetch stats data for season {season} | team {self.name} | id {self.team_id}: {e}")
+
+        # Store the fetched stats data
+        self.stats[season] = all_stats
 
     def get_team_colors(self) -> dict:
         """
@@ -224,15 +259,27 @@ class Team:
             seasonal game results.
 
         """
-
         api_info = lookup_league_api_info(league_abbv=self.espn_instance.league_abbv)
         url = f'http://sports.core.api.espn.com/{v}/sports/{api_info["sport"]}/leagues/{api_info["league"]}/seasons/{season}/types/2/teams/{self.team_id}/record?lang=en&region=us'
-        results_content = fetch_espn_data(url)
         season_records = []
-        for result in results_content.get('items', []):
-            season_records.append(Record(record_json=result,
-                                         espn_instance=self.espn_instance))
-        self.records[season] = season_records
+
+        try:
+            with ThreadPoolExecutor(max_workers=1) as executor:
+                future = executor.submit(fetch_espn_data, url)
+                results_content = future.result(timeout=10)  # timeout in seconds
+
+                for result in results_content.get('items', []):
+                    season_records.append(Record(record_json=result,
+                                                 espn_instance=self.espn_instance))
+
+            self.records[season] = season_records
+
+        except TimeoutError:
+            print(f"Timeout while fetching season records for season {season} | team {self.name} | id {self.team_id}")
+        except API400Error as e:
+            print(f"API error while fetching season records for season {season} | team {self.name} | id {self.team_id}: {e}")
+        except Exception as e:
+            print(f"Unexpected error while fetching season records for season {season} | team {self.name} | id {self.team_id}: {e}")
 
     def load_season_coaches(self, season):
         """
@@ -268,6 +315,46 @@ class Team:
                                         espn_instance=self.espn_instance))
 
         self.coaches[season] = coach_records
+
+    def load_season_betting_records(self, season):
+        """
+        Fetches and loads the betting odds records for a specific team and season using concurrent requests.
+
+        This method queries the ESPN API for team-specific betting records for the given season. It handles
+        pagination and leverages a thread pool to fetch data pages concurrently, improving performance. The
+        fetched data is parsed into `Record` instances and stored in the `self.betting` dictionary under
+        the specified season.
+
+        Args:
+            season (int): The season year to fetch betting odds records for.
+
+        Raises:
+            API400Error: If the ESPN API returns a 400 error (e.g., invalid season or unavailable data).
+        """
+        api_info = lookup_league_api_info(league_abbv=self.espn_instance.league_abbv)
+        futures = []
+        url = f'http://sports.core.api.espn.com/{v}/sports/{api_info["sport"]}/leagues/{api_info["league"]}/seasons/{season}/types/0/teams/{self.team_id}/odds-records'
+
+        try:
+            season_content = fetch_espn_data(url)
+            pages = season_content.get('pageCount', 0)
+
+            with ThreadPoolExecutor() as executor:
+                future_to_page = {
+                    executor.submit(fetch_espn_data, f'{url}?page={page}'): page
+                    for page in range(1, pages + 1)
+                }
+
+                for future in as_completed(future_to_page):
+                    page_data = future.result()
+                    for bet in page_data.get('items', []):
+                        futures.append(Record(record_json=bet,
+                                              espn_instance=self.espn_instance))
+
+            self.betting[season] = futures
+
+        except API400Error as e:
+            print(f"Failed to fetch oddsbetting data for season {season} | team {self.name} | id {self.team_id}: {e}")
 
     def to_dict(self) -> dict:
         """