+++ pyespn/data/data_import.py
@@ -55,3 +55,6 @@ nascar_teams_data = open_json(team_lookup_file)
 
 team_lookup_file = 'files/indy_teams_lookup.json'
 indy_teams_data = open_json(team_lookup_file)
+
+team_lookup_file = 'files/epl_teams_ints.json'
+epl_teams_data = open_json(team_lookup_file)