+++ tests/manual/test.py
@@ -8,8 +8,9 @@ pick_round = 2
 if __name__ == '__main__':
 
     espn = PYESPN('nfl')
-    espn.load_season_coaches(season=season)
-    espn.load_season_rosters(season=2025)
+    #espn.load_standings(season=season)
+    #espn.load_regular_season_schedule(season=season)
+    espn.load_seasons_betting_records(season=2024)
     #stats = espn.get_players_historical_stats(player_id=player_id)
     #espn.load_athletes(season=season)
     #espn.load_regular_season_schedule(season=season)