+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.3.2
+* added in epl
+* added in pulling betting records for a season for teams/ teams in a league
+* fixed pulling team stats
+* removed client functions that are now found within other classes. i.e. betting calls are available in teams
+* fixed some functions to use futures, this probably wont improve much performance but if a call hangs it could help
+* added func to load futures (betting) for league class for a season
+* 
+
 ## 0.3.1
 * added function to client/team to pull a seasons roster
 * adding in tests and fixed some of the long running tests